backend/requirements.txt
awscrt==0.23.5
AWSIoTPythonSDK==1.5.4
awsiotsdk @ file:///C:/Users/admin/Desktop/ECOBIN/aws-iot-device-sdk-python-v2
boto3==1.37.38
botocore==1.37.38
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.5
six==1.17.0
urllib3==2.4.0
